Maintenance
* upgrades go EOS lib
* gets demo contract in sync with the latest eosio.cdt version (1.6.0)
* upgrades local node to the latest version (1.7.3)
* get local setup aligned with other projects (build, packaging,
  compose)
* updates helm chart

+++ storage/eos/resources.go
@@ -37,7 +37,7 @@ func parseAvailableResources(response *eos.AccountResp) *resources {
 	// Get net
 	out.net = int64(response.NetLimit.Available)
 	// Get ram
-	out.ram = response.RAMQuota - response.RAMUsage
+	out.ram = int64(response.RAMQuota - response.RAMUsage)
 
 	return out
 }
